+++ src/extension.ts
@@ -192,10 +192,12 @@ function copyResults(provider: TreeDataProvider): void {
 	let resultString: string = "";
 	const results: FindResult[] = provider.getResults();
 	for (const result of results) {
-		if ((result.line === undefined) || (result.columnBegin === undefined) || (result.columnEnd === undefined)) {
+		if ((result.line === undefined) || (result.columnBegin === undefined) || (result.columnEnd === undefined) ||
+			(result.index === undefined)) {
 			resultString += `${result.text}\n`;
 		} else {
-			resultString += `${result.line + 1}:${result.columnBegin + 1}\t${result.text}\n`;
+			resultString +=
+				`${result.index}\t${result.line + 1}:${result.columnBegin + 1}-${result.columnEnd}\t${result.text}\n`;
 		}
 	}