+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+- Add find index to results.
+
 ## [0.6.0] - 2020-05-27
 
 - Add feature to copy results to clipboard